major refactor of codebase to separate into: - scraper, which scrapes articles and stores them in a data class - other document features: stores document classes - document edits: functions that format and add to a word document
