+++ news_data_scraper/scraper/bu_tags.py
@@ -1,0 +1,3 @@
+class BuTag:
+    ER_RN = "ER-RN"
+    ER_CGA = "ER-CGA"
