+++ news_data_scraper/scraper/scraper_classes.py
@@ -1,0 +1,21 @@
+from dataclasses import dataclass, field
+from typing import List
+
+from news_data_scraper.scraper.bu_tags import BuTag
+
+
+@dataclass
+class NewsArticle:
+    """Class for scraped news article"""
+    bu_tag: BuTag
+    title: str = "Title"
+    source: str = "Source"
+    date: str = "Date"
+    text: str = "Text"
+    url: str = "Url"
+
+
+@dataclass
+class AllNewsArticles:
+    """Class containing all scraped news articles"""
+    articles: List[NewsArticle] = field(default_factory=list)
